add script to rename nuxt's output files
Nuxt adds an underline prefix before the directory name of the generated
assets and before the payload JSON. GitHub returns a 404  http  response
to files whose names start with underline, therefore it  does  not  load
script files built by nuxt. The solution is to rename  those  files  and
any reference to them.